requirements.txt
line-bot-sdk==3.7.0
flask
cloudinary
BeautifulSoup4
lxml
opencc-python-reimplemented


